+++ Samples/GDScriptIntegration/SimpleGDScriptOptionsPresenter.gd
@@ -1,67 +1,42 @@
-extends Node
 # Example of writing an options presenter in GDScript
+extends Node
+
 @export var option_item_prefab : PackedScene
 @export var options_container: Container
-@export var presenter_control: Control 
-
 var option_selected_handler: Callable 
-
 var option_selected : bool = false 
+
 func _ready() -> void: 
-	presenter_control.visible = false 
-	
-# Example options array: 
-#[
-	#{
-		#"dialogue_option_id": 0,
-		#"is_available": true,
-		#"line": {
-			#"metadata": [],
-			#"text": {
-				#"attributes": [],
-				#"text": "Yes",
-				#"text_without_character_name": "Yes"
-			#}
-		#},
-		#"text_id": "line:b7aaff9b"
-	#},
-	#{
-		#"dialogue_option_id": 1,
-		#"is_available": true,
-		#"line": {
-			#"metadata": [],
-			#"text": {
-				#"attributes": [],
-				#"text": "No",
-				#"text_without_character_name": "No"
-			#}
-		#},
-		#"text_id": "line:9bcbf175"
-	#}
-#]
+	options_container.visible = false 
+
 func run_options_async(options: Array, on_option_selected: Callable) -> void:
 	print("Options: %s"  % JSON.stringify(options))
-	option_selected = false
+
+	# options is a Dictionary converted from the LocalizedLine C# Class
+	# converts to GDScript DialogueOptions array
+	var dialogue_options := YarnSpinner.dialogue_options_from_array(options)
+	await _run_options_internal(dialogue_options, on_option_selected)
+
+func _run_options_internal(dialogue_options: Array[YarnSpinner.DialogueOption], on_option_selected: Callable) -> void:
 	# You can do await statements here if you want.
 	await get_tree().process_frame
-	option_selected_handler = on_option_selected
-	while options_container.get_child_count() >0:
-		options_container.remove_child(options_container.get_child(0))
-	for option in options:
-		if not option["is_available"]:
-			# don't render unvailable options
-			continue 
-		var option_item: SimpleGDScriptOptionItem = option_item_prefab.instantiate() 
-		option_item.set_option(option, select_option)
-		options_container.add_child(option_item)
-	
-	presenter_control.visible = true 
-	while not option_selected:
-		await get_tree().process_frame
-	
-func select_option(option_id: int) -> void:
-	option_selected_handler.call(option_id)
-	presenter_control.visible = false 
+	option_selected = false
 	while options_container.get_child_count() > 0:
 		options_container.remove_child(options_container.get_child(0))
-	option_selected = true
+
+	for option in dialogue_options:
+		var option_item: Button = option_item_prefab.instantiate() 
+		
+		option_item.disabled = !option.is_available # disable unavaliable options
+		option_item.text = option.line.text.text
+		option_item.pressed.connect(func (): 
+			on_option_selected.call(option.dialogue_option_id)
+			option_selected = true
+		)
+		
+		options_container.add_child(option_item)
+	
+	options_container.visible = true 
+	while not option_selected:
+		await get_tree().process_frame
+	options_container.visible = false